requirements.txt
Flask==0.12
Jinja2==2.9.4
MarkupSafe==0.23
MySQL-python==1.2.3
Werkzeug==0.11.15
babel==2.3.4
blinker==1.4
click==6.7
coverage==4.3.4
decorator==4.0.11
flask-babel==0.11.1
flask-login==0.4.0
flask-mail==0.9.1
flask-openid==1.2.5
flask-sqlalchemy==2.1
flask-whooshalchemy==0.56
flask-wtf==0.14.2
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
pbr==1.10.0
python-openid==2.2.5
pytz==2016.10
six==1.10.0
sqlalchemy==1.1.5
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
tempita==0.5.3dev
virtualenv==15.0.3
whoosh==2.7.4
wtforms==2.1
